backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
loguru==0.7.2
pydantic-settings==2.12.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11